refactor: database 패키지 구조 개선
database 하위 도메인에 따른 패키지를 추가함.
연관된 클래스는 user, question, reply 각 패키지로 이동함.
