+++ src/main/java/com/woowa/config/HandlerConfig.java
@@ -1,8 +1,8 @@
 package com.woowa.config;
 
-import com.woowa.database.QuestionDatabase;
-import com.woowa.database.ReplyDatabase;
-import com.woowa.database.UserDatabase;
+import com.woowa.database.question.QuestionDatabase;
+import com.woowa.database.reply.ReplyDatabase;
+import com.woowa.database.user.UserDatabase;
 import com.woowa.framework.Bean;
 import com.woowa.handler.ReplyHandler;
 import com.woowa.handler.LoginHandler;
