+++ src/main/java/com/woowa/database/question/QuestionMemoryDatabase.java
@@ -1,5 +1,6 @@
-package com.woowa.database;
+package com.woowa.database.question;
 
+import com.woowa.database.Page;
 import com.woowa.model.Question;
 import com.woowa.model.Reply;
 import java.util.List;
