+++ src/main/java/com/woowa/handler/QuestionHandler.java
@@ -1,8 +1,8 @@
 package com.woowa.handler;
 
 import com.woowa.database.Page;
-import com.woowa.database.QuestionDatabase;
-import com.woowa.database.UserDatabase;
+import com.woowa.database.question.QuestionDatabase;
+import com.woowa.database.user.UserDatabase;
 import com.woowa.framework.web.HttpMethod;
 import com.woowa.framework.web.RequestMapping;
 import com.woowa.framework.web.RequestParameter;
