+++ src/main/java/com/woowa/handler/ReplyHandler.java
@@ -1,9 +1,9 @@
 package com.woowa.handler;
 
 import com.woowa.database.Page;
-import com.woowa.database.QuestionDatabase;
-import com.woowa.database.ReplyDatabase;
-import com.woowa.database.UserDatabase;
+import com.woowa.database.question.QuestionDatabase;
+import com.woowa.database.reply.ReplyDatabase;
+import com.woowa.database.user.UserDatabase;
 import com.woowa.framework.web.HttpMethod;
 import com.woowa.framework.web.RequestMapping;
 import com.woowa.framework.web.ResponseEntity;
