+++ src/test/java/com/woowa/servlet/QuestionsServletTest.java
@@ -2,10 +2,10 @@ package com.woowa.servlet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.woowa.database.QuestionDatabase;
-import com.woowa.database.QuestionMemoryDatabase;
-import com.woowa.database.UserDatabase;
-import com.woowa.database.UserMemoryDatabase;
+import com.woowa.database.question.QuestionDatabase;
+import com.woowa.database.question.QuestionMemoryDatabase;
+import com.woowa.database.user.UserDatabase;
+import com.woowa.database.user.UserMemoryDatabase;
 import com.woowa.handler.QuestionHandler;
 import com.woowa.model.User;
 import com.woowa.support.StubHttpServletRequest;
