+++ src/test/java/com/woowa/servlet/UserProfileServletTest.java
@@ -2,8 +2,8 @@ package com.woowa.servlet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.woowa.database.UserDatabase;
-import com.woowa.database.UserMemoryDatabase;
+import com.woowa.database.user.UserDatabase;
+import com.woowa.database.user.UserMemoryDatabase;
 import com.woowa.handler.UserHandler;
 import com.woowa.model.User;
 import com.woowa.support.StubHttpServletRequest;
